+++ instrumentation-genai/opentelemetry-instrumentation-vertexai/tests/test_function_calling.py
@@ -44,10 +44,8 @@ def test_function_call_choice(
     logs = log_exporter.get_finished_logs()
     assert len(logs) == 2
     user_log, choice_log = [log_data.log_record for log_data in logs]
-    assert user_log.attributes == {
-        "gen_ai.system": "vertex_ai",
-        "event.name": "gen_ai.user.message",
-    }
+    assert user_log.attributes == {"gen_ai.system": "vertex_ai"}
+    assert user_log.event_name == "gen_ai.user.message"
     assert user_log.body == {
         "content": [
             {"text": "Get weather details in New Delhi and San Francisco?"}
@@ -55,10 +53,8 @@ def test_function_call_choice(
         "role": "user",
     }
 
-    assert choice_log.attributes == {
-        "gen_ai.system": "vertex_ai",
-        "event.name": "gen_ai.choice",
-    }
+    assert choice_log.attributes == {"gen_ai.system": "vertex_ai"}
+    assert choice_log.event_name == "gen_ai.choice"
     assert choice_log.body == {
         "finish_reason": "stop",
         "index": 0,
@@ -111,18 +107,14 @@ def test_function_call_choice_no_content(
     logs = log_exporter.get_finished_logs()
     assert len(logs) == 2
     user_log, choice_log = [log_data.log_record for log_data in logs]
-    assert user_log.attributes == {
-        "gen_ai.system": "vertex_ai",
-        "event.name": "gen_ai.user.message",
-    }
+    assert user_log.attributes == {"gen_ai.system": "vertex_ai"}
+    assert user_log.event_name == "gen_ai.user.message"
     assert user_log.body == {
         "role": "user",
     }
 
-    assert choice_log.attributes == {
-        "gen_ai.system": "vertex_ai",
-        "event.name": "gen_ai.choice",
-    }
+    assert choice_log.attributes == {"gen_ai.system": "vertex_ai"}
+    assert choice_log.event_name == "gen_ai.choice"
     assert choice_log.body == {
         "finish_reason": "stop",
         "index": 0,
@@ -172,10 +164,8 @@ def test_tool_events(
     user_log, assistant_log, tool_log1, tool_log2, choice_log = [
         log_data.log_record for log_data in logs
     ]
-    assert user_log.attributes == {
-        "gen_ai.system": "vertex_ai",
-        "event.name": "gen_ai.user.message",
-    }
+    assert user_log.attributes == {"gen_ai.system": "vertex_ai"}
+    assert user_log.event_name == "gen_ai.user.message"
     assert user_log.body == {
         "content": [
             {"text": "Get weather details in New Delhi and San Francisco?"}
@@ -183,10 +173,8 @@ def test_tool_events(
         "role": "user",
     }
 
-    assert assistant_log.attributes == {
-        "gen_ai.system": "vertex_ai",
-        "event.name": "gen_ai.assistant.message",
-    }
+    assert assistant_log.attributes == {"gen_ai.system": "vertex_ai"}
+    assert assistant_log.event_name == "gen_ai.assistant.message"
     assert assistant_log.body == {
         "role": "model",
         "content": [
@@ -205,10 +193,8 @@ def test_tool_events(
         ],
     }
 
-    assert tool_log1.attributes == {
-        "gen_ai.system": "vertex_ai",
-        "event.name": "gen_ai.tool.message",
-    }
+    assert tool_log1.attributes == {"gen_ai.system": "vertex_ai"}
+    assert tool_log1.event_name == "gen_ai.tool.message"
 
     assert tool_log1.body == {
         "role": "user",
@@ -216,20 +202,16 @@ def test_tool_events(
         "content": {"content": '{"temperature": 35, "unit": "C"}'},
     }
 
-    assert tool_log2.attributes == {
-        "gen_ai.system": "vertex_ai",
-        "event.name": "gen_ai.tool.message",
-    }
+    assert tool_log2.attributes == {"gen_ai.system": "vertex_ai"}
+    assert tool_log2.event_name == "gen_ai.tool.message"
     assert tool_log2.body == {
         "role": "user",
         "id": "get_current_weather_1",
         "content": {"content": '{"temperature": 25, "unit": "C"}'},
     }
 
-    assert choice_log.attributes == {
-        "gen_ai.system": "vertex_ai",
-        "event.name": "gen_ai.choice",
-    }
+    assert choice_log.attributes == {"gen_ai.system": "vertex_ai"}
+    assert choice_log.event_name == "gen_ai.choice"
     assert choice_log.body == {
         "finish_reason": "stop",
         "index": 0,
@@ -274,34 +256,25 @@ def test_tool_events_no_content(
     user_log, assistant_log, tool_log1, tool_log2, choice_log = [
         log_data.log_record for log_data in logs
     ]
-    assert user_log.attributes == {
-        "gen_ai.system": "vertex_ai",
-        "event.name": "gen_ai.user.message",
-    }
+    assert user_log.attributes == {"gen_ai.system": "vertex_ai"}
+    assert user_log.event_name == "gen_ai.user.message"
     assert user_log.body == {"role": "user"}
 
-    assert assistant_log.attributes == {
-        "gen_ai.system": "vertex_ai",
-        "event.name": "gen_ai.assistant.message",
-    }
+    assert assistant_log.attributes == {"gen_ai.system": "vertex_ai"}
+    assert assistant_log.event_name == "gen_ai.assistant.message"
     assert assistant_log.body == {"role": "model"}
 
-    assert tool_log1.attributes == {
-        "gen_ai.system": "vertex_ai",
-        "event.name": "gen_ai.tool.message",
-    }
+    assert tool_log1.attributes == {"gen_ai.system": "vertex_ai"}
+    assert tool_log1.event_name == "gen_ai.tool.message"
+
     assert tool_log1.body == {"role": "user", "id": "get_current_weather_0"}
 
-    assert tool_log2.attributes == {
-        "gen_ai.system": "vertex_ai",
-        "event.name": "gen_ai.tool.message",
-    }
+    assert tool_log2.attributes == {"gen_ai.system": "vertex_ai"}
+    assert tool_log2.event_name == "gen_ai.tool.message"
     assert tool_log2.body == {"role": "user", "id": "get_current_weather_1"}
 
-    assert choice_log.attributes == {
-        "gen_ai.system": "vertex_ai",
-        "event.name": "gen_ai.choice",
-    }
+    assert choice_log.attributes == {"gen_ai.system": "vertex_ai"}
+    assert choice_log.event_name == "gen_ai.choice"
     assert choice_log.body == {
         "finish_reason": "stop",
         "index": 0,